extract SupportsDelegationToken
